chore: Update application layout and routes
The changes in this commit update the application layout and routes.

In the `application_controller.rb` file, the `ActiveStorage::SetCurrent` module is included to set current URL options for active storage.

In the `application.html.erb` file, the `main` element's `flex` class is removed from the `container` class.

In the `routes.rb` file, the root path is changed to `home#index` and a new resource `products` is added with only `index` and `show` actions.

These changes improve the application's layout and routing.